folder structure refactored.
